+++ app/src/main/java/com/belkanoid/secretchat2/domain/model/Queue.kt
@@ -3,8 +3,8 @@ package com.belkanoid.secretchat2.domain.model
 import com.google.gson.annotations.SerializedName
 
 data class Queue(
-    val id: Int,
-    val userId: Int,
-    val messageId: Int,
-    val timestamp: Int
+    val id: Long,
+    val userId: Long,
+    val messageId: Long,
+    val timestamp: Long
 )
